requirements.txt
fastapi==0.59.0
gensim==3.8.3
langdetect==1.0.8
matplotlib==3.2.2
matplotlib-inline==0.1.2
nbconvert==5.6.1
nltk==3.6.2
numpy==1.19.5
pandas==1.0.5
pyenchant==3.2.1
pyspellchecker==0.6.2
pyLDAvis==3.2.2
PyYAML==5.4.1
regex==2021.4.4
scikit-learn==0.23.1
sentistrength==0.0.9
spacy==2.3.5
streamlit==0.64.0
tokenizers==0.8.1rc1
torch==1.8.1
tqdm==4.61.0
transformers==3.0.2
uvicorn==0.11.7